Break up reference models
